Consolidate Vm implementations
As much as the two vm implementations SEEM different, the differences
largely lie in how we're getting the nbd proxy socket.  One is relying
on launching a process (nbd-proxy), the other is getting the fd from
dbus.  Given [1] exists and is in process, we need to have a plan for
getting these two VM implementations into one, once that patchset is
complete.

This commit: Splits the vm-websocket option into vm-websocket-provider,
providing two options, nbd-proxy, and virtual-media (the names of the
respective apps).  To accomplish this, it moves the contents of
nbd-proxy into include/vm-websocket, so we can compare the similarities
and start consolidating.

The longer term intent is that the nbd-proxy option will be completely
removed, and the code deleted.  This has the additional advantage that
we will no longer require the boost::process dependency, as all info
will be available on dbus.

As part of this, the nbd proxy websocket is also registered at /vm/0/0,
to be backward compatible with the old interfaces.

Tested: Code compiles.  Need some help here.

[1] https://gerrit.openbmc.org/c/openbmc/jsnbd/+/49944
