+++ meson.build
@@ -94,7 +94,6 @@ feature_map = {
     'static-hosting': '-DBMCWEB_ENABLE_STATIC_HOSTING',
     'experimental-redfish-multi-computer-system': '-DBMCWEB_ENABLE_MULTI_COMPUTERSYSTEM',
     'redfish-license': '-DBMCWEB_ENABLE_REDFISH_LICENSE',
-    'vm-websocket': '-DBMCWEB_ENABLE_VM_WEBSOCKET',
     'xtoken-auth': '-DBMCWEB_ENABLE_XTOKEN_AUTHENTICATION',
     'audit-events': '-DBMCWEB_ENABLE_LINUX_AUDIT_EVENTS',
     #'vm-nbdproxy': '-DBMCWEB_ENABLE_VM_NBDPROXY',
